Extract utils from QueryIdCachingProxyHandler
