+++ gateway-ha/src/main/java/io/trino/gateway/ha/handler/QueryIdCachingProxyHandler.java
@@ -14,7 +14,6 @@
 package io.trino.gateway.ha.handler;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.google.common.base.Splitter;
 import com.google.common.collect.ImmutableList;
 import com.google.common.io.CharStreams;
 import io.airlift.log.Logger;
@@ -22,10 +21,8 @@ import io.trino.gateway.ha.config.GatewayCookieConfigurationPropertiesProvider;
 import io.trino.gateway.ha.router.GatewayCookie;
 import io.trino.gateway.ha.router.OAuth2GatewayCookie;
 import io.trino.gateway.ha.router.QueryHistoryManager;
-import io.trino.gateway.ha.router.RoutingGroupSelector;
 import io.trino.gateway.ha.router.RoutingManager;
 import io.trino.gateway.proxyserver.ProxyHandler;
-import io.trino.gateway.proxyserver.wrapper.MultiReadHttpServletRequest;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -38,14 +35,12 @@ import java.io.IOException;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.Arrays;
-import java.util.Base64;
 import java.util.HashMap;
 import java.util.List;
-import java.util.Optional;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 import static com.google.common.base.Strings.isNullOrEmpty;
+import static io.trino.gateway.ha.handler.ProxyUtils.buildUriWithNewBackend;
+import static io.trino.gateway.ha.handler.ProxyUtils.getQueryDetailsFromRequest;
 
 public class QueryIdCachingProxyHandler
         extends ProxyHandler
@@ -62,90 +57,32 @@ public class QueryIdCachingProxyHandler
     public static final String OAUTH_PATH = "/oauth2";
     public static final String AUTHORIZATION = "Authorization";
     public static final String USER_HEADER = "X-Trino-User";
-    public static final String SOURCE_HEADER = "X-Trino-Source";
     public static final String HOST_HEADER = "Host";
-    private static final int QUERY_TEXT_LENGTH_FOR_HISTORY = 200;
-    /**
-     * This regular expression matches query ids as they appear in the path of a URL. The query id must be preceded
-     * by a "/". A query id is defined as three groups of digits separated by underscores, with a final group
-     * consisting of any alphanumeric characters.
-     */
-    private static final Pattern QUERY_ID_PATH_PATTERN = Pattern.compile(".*/(\\d+_\\d+_\\d+_\\w+).*");
-    /**
-     * This regular expression matches query ids as they appear in the query parameters of a URL. The query id is
-     * defined as in QUERY_TEXT_LENGTH_FOR_HISTORY. The query id must either be at the beginning of the query parameter
-     * string, or be preceded by %2F (a URL-encoded "/"), or  "query_id=", with or without the underscore and any
-     * capitalization.
-     */
-    private static final Pattern QUERY_ID_PARAM_PATTERN = Pattern.compile(".*(?:%2F|(?i)query_?id(?-i)=|^)(\\d+_\\d+_\\d+_\\w+).*");
-
-    private static final Pattern EXTRACT_BETWEEN_SINGLE_QUOTES = Pattern.compile("'([^\\s']+)'");
 
     private static final ObjectMapper OBJECT_MAPPER = new ObjectMapper();
     private static final Logger log = Logger.get(QueryIdCachingProxyHandler.class);
 
     private final RoutingManager routingManager;
-    private final RoutingGroupSelector routingGroupSelector;
     private final QueryHistoryManager queryHistoryManager;
+    private final RoutingTargetHandler routingTargetHandler;
 
     private final ProxyHandlerStats proxyHandlerStats;
-    private final List<String> extraWhitelistPaths;
     private final String applicationEndpoint;
     private final boolean cookiesEnabled;
 
     public QueryIdCachingProxyHandler(
             QueryHistoryManager queryHistoryManager,
             RoutingManager routingManager,
-            RoutingGroupSelector routingGroupSelector,
+            RoutingTargetHandler routingTargetHandler,
             int serverApplicationPort,
-            ProxyHandlerStats proxyHandlerStats,
-            List<String> extraWhitelistPaths)
+            ProxyHandlerStats proxyHandlerStats)
     {
         this.proxyHandlerStats = proxyHandlerStats;
         this.routingManager = routingManager;
-        this.routingGroupSelector = routingGroupSelector;
         this.queryHistoryManager = queryHistoryManager;
-        this.extraWhitelistPaths = extraWhitelistPaths;
+        this.routingTargetHandler = routingTargetHandler;
         this.applicationEndpoint = "http://localhost:" + serverApplicationPort;
         cookiesEnabled = GatewayCookieConfigurationPropertiesProvider.getInstance().isEnabled();
-    }
-
-    protected static String extractQueryIdIfPresent(String path, String queryParams)
-    {
-        if (path == null) {
-            return null;
-        }
-        String queryId = null;
-
-        log.debug("trying to extract query id from  path [%s] or queryString [%s]", path, queryParams);
-        if (path.startsWith(V1_STATEMENT_PATH) || path.startsWith(V1_QUERY_PATH)) {
-            String[] tokens = path.split("/");
-            if (tokens.length >= 4) {
-                if (path.contains("queued")
-                        || path.contains("scheduled")
-                        || path.contains("executing")
-                        || path.contains("partialCancel")) {
-                    queryId = tokens[4];
-                }
-                else {
-                    queryId = tokens[3];
-                }
-            }
-        }
-        else if (path.startsWith(TRINO_UI_PATH)) {
-            Matcher matcher = QUERY_ID_PATH_PATTERN.matcher(path);
-            if (matcher.matches()) {
-                queryId = matcher.group(1);
-            }
-        }
-        if (!isNullOrEmpty(queryParams)) {
-            Matcher matcher = QUERY_ID_PARAM_PATTERN.matcher(queryParams);
-            if (matcher.matches()) {
-                queryId = matcher.group(1);
-            }
-        }
-        log.debug("query id in url [%s]", queryId);
-        return queryId;
     }
 
     static void setForwardedHostHeaderOnProxyRequest(HttpServletRequest request,
@@ -174,74 +111,6 @@ public class QueryIdCachingProxyHandler
         }
     }
 
-    static String getQueryUser(HttpServletRequest request)
-    {
-        String trinoUser = request.getHeader(USER_HEADER);
-
-        if (!isNullOrEmpty(trinoUser)) {
-            log.info("user from %s", USER_HEADER);
-            return trinoUser;
-        }
-
-        log.info("user from basic auth");
-        String user = "";
-        String header = request.getHeader(AUTHORIZATION);
-        if (header == null) {
-            log.error("didn't find any basic auth header");
-            return user;
-        }
-
-        int space = header.indexOf(' ');
-        if ((space < 0) || !header.substring(0, space).equalsIgnoreCase("basic")) {
-            log.error("basic auth format is incorrect");
-            return user;
-        }
-
-        String headerInfo = header.substring(space + 1).trim();
-        if (isNullOrEmpty(headerInfo)) {
-            log.error("The encoded value of basic auth doesn't exist");
-            return user;
-        }
-
-        String info = new String(Base64.getDecoder().decode(headerInfo));
-        List<String> parts = Splitter.on(':').limit(2).splitToList(info);
-        if (parts.size() < 1) {
-            log.error("no user inside the basic auth text");
-            return user;
-        }
-        return parts.get(0);
-    }
-
-    protected String extractQueryIdIfPresent(HttpServletRequest request)
-    {
-        String path = request.getRequestURI();
-        String queryParams = request.getQueryString();
-        try {
-            String queryText = CharStreams.toString(request.getReader());
-            if (!isNullOrEmpty(queryText)
-                    && queryText.toLowerCase().contains("system.runtime.kill_query")) {
-                // extract and return the queryId
-                String[] parts = queryText.split(",");
-                for (String part : parts) {
-                    if (part.contains("query_id")) {
-                        Matcher m = EXTRACT_BETWEEN_SINGLE_QUOTES.matcher(part);
-                        if (m.find()) {
-                            String queryQuoted = m.group();
-                            if (!isNullOrEmpty(queryQuoted) && queryQuoted.length() > 0) {
-                                return queryQuoted.substring(1, queryQuoted.length() - 1);
-                            }
-                        }
-                    }
-                }
-            }
-        }
-        catch (Exception e) {
-            log.error(e, "Error extracting query payload from request");
-        }
-
-        return extractQueryIdIfPresent(path, queryParams);
-    }
-
     @Override
     public void preConnectionHook(HttpServletRequest request, Request proxyRequest)
     {
@@ -261,21 +130,9 @@ public class QueryIdCachingProxyHandler
             }
         }
 
-        if (isPathWhiteListed(request.getRequestURI())) {
+        if (routingTargetHandler.isPathWhiteListed(request.getRequestURI())) {
             setForwardedHostHeaderOnProxyRequest(request, proxyRequest);
         }
-    }
-
-    private boolean isPathWhiteListed(String path)
-    {
-        return path.startsWith(V1_STATEMENT_PATH)
-                || path.startsWith(V1_QUERY_PATH)
-                || path.startsWith(TRINO_UI_PATH)
-                || path.startsWith(V1_INFO_PATH)
-                || path.startsWith(V1_NODE_PATH)
-                || path.startsWith(UI_API_STATS_PATH)
-                || path.startsWith(OAUTH_PATH)
-                || extraWhitelistPaths.stream().anyMatch(s -> path.startsWith(s));
     }
 
     @Override
@@ -300,67 +157,10 @@ public class QueryIdCachingProxyHandler
     @Override
     public String rewriteTarget(HttpServletRequest request)
     {
-        if (!isPathWhiteListed(request.getRequestURI())) {
+        if (!routingTargetHandler.isPathWhiteListed(request.getRequestURI())) {
             return buildUriWithNewBackend(applicationEndpoint, request);
         }
-
-        Optional<String> previousBackend = getPreviousBackend(request);
-        String clusterHost = previousBackend.orElseGet(() -> getBackendFromRoutingGroup(request));
-        // set target clusterHost so that we could save queryId to cluster mapping later.
-        ((MultiReadHttpServletRequest) request).addHeader(PROXY_TARGET_HEADER, clusterHost);
-        logRewrite(clusterHost, request);
-
-        return buildUriWithNewBackend(clusterHost, request);
-    }
-
-    private void logRewrite(String newBackend, HttpServletRequest request)
-    {
-        log.info("Rerouting [%s://%s:%s%s%s]--> [%s]",
-                request.getScheme(),
-                request.getRemoteHost(),
-                request.getServerPort(),
-                request.getRequestURI(),
-                (request.getQueryString() != null ? "?" + request.getQueryString() : ""),
-                buildUriWithNewBackend(newBackend, request));
-    }
-
-    private String buildUriWithNewBackend(String backendHost, HttpServletRequest request)
-    {
-        return backendHost + request.getRequestURI() + (request.getQueryString() != null ? "?" + request.getQueryString() : "");
-    }
-
-    private String getBackendFromRoutingGroup(HttpServletRequest request)
-    {
-        String routingGroup = routingGroupSelector.findRoutingGroup(request);
-        String user = request.getHeader(USER_HEADER);
-        if (!isNullOrEmpty(routingGroup)) {
-            // This falls back on adhoc backend if there are no cluster found for the routing group.
-            return routingManager.provideBackendForRoutingGroup(routingGroup, user);
-        }
-        return routingManager.provideAdhocBackend(user);
-    }
-
-    private Optional<String> getPreviousBackend(HttpServletRequest request)
-    {
-        String queryId = extractQueryIdIfPresent(request);
-        if (!isNullOrEmpty(queryId)) {
-            return Optional.of(routingManager.findBackendForQueryId(queryId));
-        }
-        if (cookiesEnabled && request.getCookies() != null) {
-            List<GatewayCookie> cookies = Arrays.stream(request.getCookies())
-                    .filter(c -> c.getName().startsWith(GatewayCookie.PREFIX))
-                    .map(GatewayCookie::fromCookie)
-                    .filter(GatewayCookie::isValid)
-                    .filter(c -> !isNullOrEmpty(c.getBackend()))
-                    .filter(c -> c.matchesRoutingPath(request.getRequestURI()))
-                    .sorted()
-                    .toList();
-            if (!cookies.isEmpty()) {
-                return Optional.of(cookies.getFirst().getBackend());
-            }
-        }
-
-        return Optional.empty();
+        return routingTargetHandler.getRoutingDestination(request);
     }
 
     @Override
@@ -389,7 +189,7 @@ public class QueryIdCachingProxyHandler
         super.postConnectionHook(request, response, buffer, offset, length, callback);
     }
 
-    void recordBackendForQueryId(HttpServletRequest request, HttpServletResponse response, byte[] buffer)
+    private void recordBackendForQueryId(HttpServletRequest request, HttpServletResponse response, byte[] buffer)
             throws IOException
     {
         String output;
@@ -407,7 +207,7 @@ public class QueryIdCachingProxyHandler
         if (queryDetail.getBackendUrl() == null) {
             log.error("Server response to request %s does not contain proxytarget header", request.getRequestURI());
         }
-        log.debug("Extracting Proxy destination : [%s] for request : [%s]", queryDetail.getBackendUrl(), request.getRequestURI());
+        log.debug("Extracting proxy destination : [%s] for request : [%s]", queryDetail.getBackendUrl(), request.getRequestURI());
 
         if (response.getStatus() == HttpStatus.OK_200) {
             HashMap<String, String> results = OBJECT_MAPPER.readValue(output, HashMap.class);
@@ -423,21 +223,5 @@ public class QueryIdCachingProxyHandler
         }
         // Save history in Trino Gateway.
         queryHistoryManager.submitQueryDetail(queryDetail);
-    }
-
-    private QueryHistoryManager.QueryDetail getQueryDetailsFromRequest(HttpServletRequest request)
-            throws IOException
-    {
-        QueryHistoryManager.QueryDetail queryDetail = new QueryHistoryManager.QueryDetail();
-        queryDetail.setBackendUrl(request.getHeader(PROXY_TARGET_HEADER));
-        queryDetail.setCaptureTime(System.currentTimeMillis());
-        queryDetail.setUser(getQueryUser(request));
-        queryDetail.setSource(request.getHeader(SOURCE_HEADER));
-        String queryText = CharStreams.toString(request.getReader());
-        queryDetail.setQueryText(
-                queryText.length() > QUERY_TEXT_LENGTH_FOR_HISTORY
-                        ? queryText.substring(0, QUERY_TEXT_LENGTH_FOR_HISTORY) + "..."
-                        : queryText);
-        return queryDetail;
     }
 }
